Add assets/images folder with country flags
This commit adds a new folder named assets/images to the project, which contains a set of country flags in SVG format. These flags will be used in the newsfeed customization menu to allow users to select a region. The flags have been resized to 32x32 pixels for better display in the menu. The new folder is included in the project's .gitignore file to prevent any sensitive information from being accidentally committed to the repository.