Feat(NewsFeed): add elements allowing user to search by regio using icons of flags.
This commit adds new options to the NewsFeed component that allow users to filter news by region, kind, currency, and type. Also refactors the code to improve readability and adds comments to explain the new features. - strongly suggest we go back to using tailwind or MUI - making it like this is just not feasable for the stylin - spent far too long on somerthing that could have otherwise been done in hours using a component