feat: Add custom currency component
Adding  new React component called CustomCurrency. Allows users to enter a custom short code (e.g. BTC) to filter news articles based on the selected currency.

The CustomCurrency component is implemented as a controlled component that updates its state as the user types into an input field. The user's input can be submitted via a button to trigger a callback function to update the
ewsOptions state in the parent component.

This commit also includes CSS styles for the CustomCurrency component in a separate file CustomCurrency.module.css. The CSS styles include styling for the header, input field, and submit button.